+++ bindings/c/include/jolt/Jolt/Geometry/AABox.h
@@ -102,31 +102,31 @@ JOLT_API void JPH_AABox_DestroyArray(const JPH_AABox *_this);
 JOLT_API JPH_AABox *JPH_AABox_AssignFromAnother(JPH_AABox *_this, const JPH_AABox *_other);
 
 /// Generated from method `JPH::AABox::operator new`.
-JOLT_API void *Jolt_new_JPH_AABox_size_t(unsigned long inCount);
+JOLT_API void *Jolt_new_JPH_AABox_size_t(size_t inCount);
 
 /// Generated from method `JPH::AABox::operator delete`.
 JOLT_API void Jolt_delete_JPH_AABox_void_ptr(void *inPointer);
 
 /// Generated from method `JPH::AABox::operator delete`.
-JOLT_API void Jolt_delete_JPH_AABox_void_ptr_size_t(void *inPointer, unsigned long inSize);
+JOLT_API void Jolt_delete_JPH_AABox_void_ptr_size_t(void *inPointer, size_t inSize);
 
 /// Generated from method `JPH::AABox::operator new[]`.
-JOLT_API void *Jolt_new_array_JPH_AABox_size_t(unsigned long inCount);
+JOLT_API void *Jolt_new_array_JPH_AABox_size_t(size_t inCount);
 
 /// Generated from method `JPH::AABox::operator delete[]`.
 JOLT_API void Jolt_delete_array_JPH_AABox_void_ptr(void *inPointer);
 
 /// Generated from method `JPH::AABox::operator delete[]`.
-JOLT_API void Jolt_delete_array_JPH_AABox_void_ptr_size_t(void *inPointer, unsigned long inSize);
+JOLT_API void Jolt_delete_array_JPH_AABox_void_ptr_size_t(void *inPointer, size_t inSize);
 
 /// Generated from method `JPH::AABox::operator new`.
-JOLT_API void *Jolt_new_JPH_AABox_size_t_void_ptr(unsigned long inCount, void *inPointer);
+JOLT_API void *Jolt_new_JPH_AABox_size_t_void_ptr(size_t inCount, void *inPointer);
 
 /// Generated from method `JPH::AABox::operator delete`.
 JOLT_API void Jolt_delete_JPH_AABox_void_ptr_void_ptr(void *inPointer, void *inPlace);
 
 /// Generated from method `JPH::AABox::operator new[]`.
-JOLT_API void *Jolt_new_array_JPH_AABox_size_t_void_ptr(unsigned long inCount, void *inPointer);
+JOLT_API void *Jolt_new_array_JPH_AABox_size_t_void_ptr(size_t inCount, void *inPointer);
 
 /// Generated from method `JPH::AABox::operator delete[]`.
 JOLT_API void Jolt_delete_array_JPH_AABox_void_ptr_void_ptr(void *inPointer, void *inPlace);
