+++ bindings/c/include/jolt/Jolt/Math/BVec16.h
@@ -68,31 +68,31 @@ JOLT_API void JPH_BVec16_DestroyArray(const JPH_BVec16 *_this);
 JOLT_API JPH_BVec16 *JPH_BVec16_AssignFromAnother(JPH_BVec16 *_this, const JPH_BVec16 *inRHS);
 
 /// Generated from method `JPH::BVec16::operator new`.
-JOLT_API void *Jolt_new_JPH_BVec16_size_t(unsigned long inCount);
+JOLT_API void *Jolt_new_JPH_BVec16_size_t(size_t inCount);
 
 /// Generated from method `JPH::BVec16::operator delete`.
 JOLT_API void Jolt_delete_JPH_BVec16_void_ptr(void *inPointer);
 
 /// Generated from method `JPH::BVec16::operator delete`.
-JOLT_API void Jolt_delete_JPH_BVec16_void_ptr_size_t(void *inPointer, unsigned long inSize);
+JOLT_API void Jolt_delete_JPH_BVec16_void_ptr_size_t(void *inPointer, size_t inSize);
 
 /// Generated from method `JPH::BVec16::operator new[]`.
-JOLT_API void *Jolt_new_array_JPH_BVec16_size_t(unsigned long inCount);
+JOLT_API void *Jolt_new_array_JPH_BVec16_size_t(size_t inCount);
 
 /// Generated from method `JPH::BVec16::operator delete[]`.
 JOLT_API void Jolt_delete_array_JPH_BVec16_void_ptr(void *inPointer);
 
 /// Generated from method `JPH::BVec16::operator delete[]`.
-JOLT_API void Jolt_delete_array_JPH_BVec16_void_ptr_size_t(void *inPointer, unsigned long inSize);
+JOLT_API void Jolt_delete_array_JPH_BVec16_void_ptr_size_t(void *inPointer, size_t inSize);
 
 /// Generated from method `JPH::BVec16::operator new`.
-JOLT_API void *Jolt_new_JPH_BVec16_size_t_void_ptr(unsigned long inCount, void *inPointer);
+JOLT_API void *Jolt_new_JPH_BVec16_size_t_void_ptr(size_t inCount, void *inPointer);
 
 /// Generated from method `JPH::BVec16::operator delete`.
 JOLT_API void Jolt_delete_JPH_BVec16_void_ptr_void_ptr(void *inPointer, void *inPlace);
 
 /// Generated from method `JPH::BVec16::operator new[]`.
-JOLT_API void *Jolt_new_array_JPH_BVec16_size_t_void_ptr(unsigned long inCount, void *inPointer);
+JOLT_API void *Jolt_new_array_JPH_BVec16_size_t_void_ptr(size_t inCount, void *inPointer);
 
 /// Generated from method `JPH::BVec16::operator delete[]`.
 JOLT_API void Jolt_delete_array_JPH_BVec16_void_ptr_void_ptr(void *inPointer, void *inPlace);
