+++ bindings/c/include/jolt/Jolt/Math/DVec3.h
@@ -98,31 +98,31 @@ JOLT_API JPH_Vec3 *JPH_DVec3_ConvertTo_JPH_Vec3(const JPH_DVec3 *_this);
 JOLT_API JPH_DVec3 *JPH_DVec3_AssignFromAnother(JPH_DVec3 *_this, const JPH_DVec3 *inRHS);
 
 /// Generated from method `JPH::DVec3::operator new`.
-JOLT_API void *Jolt_new_JPH_DVec3_size_t(unsigned long inCount);
+JOLT_API void *Jolt_new_JPH_DVec3_size_t(size_t inCount);
 
 /// Generated from method `JPH::DVec3::operator delete`.
 JOLT_API void Jolt_delete_JPH_DVec3_void_ptr(void *inPointer);
 
 /// Generated from method `JPH::DVec3::operator delete`.
-JOLT_API void Jolt_delete_JPH_DVec3_void_ptr_size_t(void *inPointer, unsigned long inSize);
+JOLT_API void Jolt_delete_JPH_DVec3_void_ptr_size_t(void *inPointer, size_t inSize);
 
 /// Generated from method `JPH::DVec3::operator new[]`.
-JOLT_API void *Jolt_new_array_JPH_DVec3_size_t(unsigned long inCount);
+JOLT_API void *Jolt_new_array_JPH_DVec3_size_t(size_t inCount);
 
 /// Generated from method `JPH::DVec3::operator delete[]`.
 JOLT_API void Jolt_delete_array_JPH_DVec3_void_ptr(void *inPointer);
 
 /// Generated from method `JPH::DVec3::operator delete[]`.
-JOLT_API void Jolt_delete_array_JPH_DVec3_void_ptr_size_t(void *inPointer, unsigned long inSize);
+JOLT_API void Jolt_delete_array_JPH_DVec3_void_ptr_size_t(void *inPointer, size_t inSize);
 
 /// Generated from method `JPH::DVec3::operator new`.
-JOLT_API void *Jolt_new_JPH_DVec3_size_t_void_ptr(unsigned long inCount, void *inPointer);
+JOLT_API void *Jolt_new_JPH_DVec3_size_t_void_ptr(size_t inCount, void *inPointer);
 
 /// Generated from method `JPH::DVec3::operator delete`.
 JOLT_API void Jolt_delete_JPH_DVec3_void_ptr_void_ptr(void *inPointer, void *inPlace);
 
 /// Generated from method `JPH::DVec3::operator new[]`.
-JOLT_API void *Jolt_new_array_JPH_DVec3_size_t_void_ptr(unsigned long inCount, void *inPointer);
+JOLT_API void *Jolt_new_array_JPH_DVec3_size_t_void_ptr(size_t inCount, void *inPointer);
 
 /// Generated from method `JPH::DVec3::operator delete[]`.
 JOLT_API void Jolt_delete_array_JPH_DVec3_void_ptr_void_ptr(void *inPointer, void *inPlace);
