+++ bindings/c/include/jolt/Jolt/Math/Double3.h
@@ -110,31 +110,31 @@ JOLT_API void JPH_Double3_DestroyArray(const JPH_Double3 *_this);
 JOLT_API JPH_Double3 *JPH_Double3_AssignFromAnother(JPH_Double3 *_this, const JPH_Double3 *inRHS);
 
 /// Generated from method `JPH::Double3::operator new`.
-JOLT_API void *Jolt_new_JPH_Double3_size_t(unsigned long inCount);
+JOLT_API void *Jolt_new_JPH_Double3_size_t(size_t inCount);
 
 /// Generated from method `JPH::Double3::operator delete`.
 JOLT_API void Jolt_delete_JPH_Double3_void_ptr(void *inPointer);
 
 /// Generated from method `JPH::Double3::operator delete`.
-JOLT_API void Jolt_delete_JPH_Double3_void_ptr_size_t(void *inPointer, unsigned long inSize);
+JOLT_API void Jolt_delete_JPH_Double3_void_ptr_size_t(void *inPointer, size_t inSize);
 
 /// Generated from method `JPH::Double3::operator new[]`.
-JOLT_API void *Jolt_new_array_JPH_Double3_size_t(unsigned long inCount);
+JOLT_API void *Jolt_new_array_JPH_Double3_size_t(size_t inCount);
 
 /// Generated from method `JPH::Double3::operator delete[]`.
 JOLT_API void Jolt_delete_array_JPH_Double3_void_ptr(void *inPointer);
 
 /// Generated from method `JPH::Double3::operator delete[]`.
-JOLT_API void Jolt_delete_array_JPH_Double3_void_ptr_size_t(void *inPointer, unsigned long inSize);
+JOLT_API void Jolt_delete_array_JPH_Double3_void_ptr_size_t(void *inPointer, size_t inSize);
 
 /// Generated from method `JPH::Double3::operator new`.
-JOLT_API void *Jolt_new_JPH_Double3_size_t_void_ptr(unsigned long inCount, void *inPointer);
+JOLT_API void *Jolt_new_JPH_Double3_size_t_void_ptr(size_t inCount, void *inPointer);
 
 /// Generated from method `JPH::Double3::operator delete`.
 JOLT_API void Jolt_delete_JPH_Double3_void_ptr_void_ptr(void *inPointer, void *inPlace);
 
 /// Generated from method `JPH::Double3::operator new[]`.
-JOLT_API void *Jolt_new_array_JPH_Double3_size_t_void_ptr(unsigned long inCount, void *inPointer);
+JOLT_API void *Jolt_new_array_JPH_Double3_size_t_void_ptr(size_t inCount, void *inPointer);
 
 /// Generated from method `JPH::Double3::operator delete[]`.
 JOLT_API void Jolt_delete_array_JPH_Double3_void_ptr_void_ptr(void *inPointer, void *inPlace);
