+++ bindings/c/include/jolt/Jolt/Math/Float2.h
@@ -95,31 +95,31 @@ JOLT_API void JPH_Float2_DestroyArray(const JPH_Float2 *_this);
 JOLT_API JPH_Float2 *JPH_Float2_AssignFromAnother(JPH_Float2 *_this, const JPH_Float2 *inRHS);
 
 /// Generated from method `JPH::Float2::operator new`.
-JOLT_API void *Jolt_new_JPH_Float2_size_t(unsigned long inCount);
+JOLT_API void *Jolt_new_JPH_Float2_size_t(size_t inCount);
 
 /// Generated from method `JPH::Float2::operator delete`.
 JOLT_API void Jolt_delete_JPH_Float2_void_ptr(void *inPointer);
 
 /// Generated from method `JPH::Float2::operator delete`.
-JOLT_API void Jolt_delete_JPH_Float2_void_ptr_size_t(void *inPointer, unsigned long inSize);
+JOLT_API void Jolt_delete_JPH_Float2_void_ptr_size_t(void *inPointer, size_t inSize);
 
 /// Generated from method `JPH::Float2::operator new[]`.
-JOLT_API void *Jolt_new_array_JPH_Float2_size_t(unsigned long inCount);
+JOLT_API void *Jolt_new_array_JPH_Float2_size_t(size_t inCount);
 
 /// Generated from method `JPH::Float2::operator delete[]`.
 JOLT_API void Jolt_delete_array_JPH_Float2_void_ptr(void *inPointer);
 
 /// Generated from method `JPH::Float2::operator delete[]`.
-JOLT_API void Jolt_delete_array_JPH_Float2_void_ptr_size_t(void *inPointer, unsigned long inSize);
+JOLT_API void Jolt_delete_array_JPH_Float2_void_ptr_size_t(void *inPointer, size_t inSize);
 
 /// Generated from method `JPH::Float2::operator new`.
-JOLT_API void *Jolt_new_JPH_Float2_size_t_void_ptr(unsigned long inCount, void *inPointer);
+JOLT_API void *Jolt_new_JPH_Float2_size_t_void_ptr(size_t inCount, void *inPointer);
 
 /// Generated from method `JPH::Float2::operator delete`.
 JOLT_API void Jolt_delete_JPH_Float2_void_ptr_void_ptr(void *inPointer, void *inPlace);
 
 /// Generated from method `JPH::Float2::operator new[]`.
-JOLT_API void *Jolt_new_array_JPH_Float2_size_t_void_ptr(unsigned long inCount, void *inPointer);
+JOLT_API void *Jolt_new_array_JPH_Float2_size_t_void_ptr(size_t inCount, void *inPointer);
 
 /// Generated from method `JPH::Float2::operator delete[]`.
 JOLT_API void Jolt_delete_array_JPH_Float2_void_ptr_void_ptr(void *inPointer, void *inPlace);
