+++ bindings/c/include/jolt/Jolt/Math/Float4.h
@@ -127,31 +127,31 @@ JOLT_API void JPH_Float4_DestroyArray(const JPH_Float4 *_this);
 JOLT_API JPH_Float4 *JPH_Float4_AssignFromAnother(JPH_Float4 *_this, const JPH_Float4 *inRHS);
 
 /// Generated from method `JPH::Float4::operator new`.
-JOLT_API void *Jolt_new_JPH_Float4_size_t(unsigned long inCount);
+JOLT_API void *Jolt_new_JPH_Float4_size_t(size_t inCount);
 
 /// Generated from method `JPH::Float4::operator delete`.
 JOLT_API void Jolt_delete_JPH_Float4_void_ptr(void *inPointer);
 
 /// Generated from method `JPH::Float4::operator delete`.
-JOLT_API void Jolt_delete_JPH_Float4_void_ptr_size_t(void *inPointer, unsigned long inSize);
+JOLT_API void Jolt_delete_JPH_Float4_void_ptr_size_t(void *inPointer, size_t inSize);
 
 /// Generated from method `JPH::Float4::operator new[]`.
-JOLT_API void *Jolt_new_array_JPH_Float4_size_t(unsigned long inCount);
+JOLT_API void *Jolt_new_array_JPH_Float4_size_t(size_t inCount);
 
 /// Generated from method `JPH::Float4::operator delete[]`.
 JOLT_API void Jolt_delete_array_JPH_Float4_void_ptr(void *inPointer);
 
 /// Generated from method `JPH::Float4::operator delete[]`.
-JOLT_API void Jolt_delete_array_JPH_Float4_void_ptr_size_t(void *inPointer, unsigned long inSize);
+JOLT_API void Jolt_delete_array_JPH_Float4_void_ptr_size_t(void *inPointer, size_t inSize);
 
 /// Generated from method `JPH::Float4::operator new`.
-JOLT_API void *Jolt_new_JPH_Float4_size_t_void_ptr(unsigned long inCount, void *inPointer);
+JOLT_API void *Jolt_new_JPH_Float4_size_t_void_ptr(size_t inCount, void *inPointer);
 
 /// Generated from method `JPH::Float4::operator delete`.
 JOLT_API void Jolt_delete_JPH_Float4_void_ptr_void_ptr(void *inPointer, void *inPlace);
 
 /// Generated from method `JPH::Float4::operator new[]`.
-JOLT_API void *Jolt_new_array_JPH_Float4_size_t_void_ptr(unsigned long inCount, void *inPointer);
+JOLT_API void *Jolt_new_array_JPH_Float4_size_t_void_ptr(size_t inCount, void *inPointer);
 
 /// Generated from method `JPH::Float4::operator delete[]`.
 JOLT_API void Jolt_delete_array_JPH_Float4_void_ptr_void_ptr(void *inPointer, void *inPlace);
